\name{MEDIPS-package}
\alias{MEDIPS-package}
\alias{MEDIPS}
\docType{package}
\title{
(MeD)IP-seq data analysis
}
\description{
MEDIPS was developed for analyzing data derived from methylated DNA immunoprecipitation (MeDIP) experiments followed by sequencing (MeDIP-seq). 
Nevertheless, several functionalities may be applied to other types of sequencing data (e.g. differential coverage or testing the saturation of ChIP-seq data). 
MEDIPS addresses several aspects in the context of MeDIP-seq data analysis including basic data processing, several quality controls, normalization, and identification of differential coverage.
}
\details{
\tabular{ll}{
Package: \tab MEDIPS\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Lukas Chavez, Matthias Lienhard, Joern Dietrich

Maintainer: Lukas Chavez <lchavez@liai.org>
}
\references{
Chavez L, Jozefczuk J, Grimm C, Dietrich J, Timmermann B, Lehrach H, Herwig R, Adjaye J., Computational analysis of genome-wide DNA methylation during the differentiation of human embryonic stem cells along the endodermal lineage, Genome Res. 2010 Oct;20(10):1441-50. Epub 2010 Aug 27.
}
\keyword{ package }
