\name{MEDIPS.plotCoverage}
\alias{MEDIPS.plotCoverage}
\title{
Function plots the results of the MEDIPS.coverageAnalysis function.
}
\description{
This function has been deprecated. 
Please see MEDIPS.plotSeqCoverage instead.

The results of the coverage analysis will be visualized by the function.
}
\usage{
MEDIPS.plotCoverage(coverageObj = NULL)
}
\arguments{
  \item{coverageObj}{
The coverage results object returned by the MEDIPS.coverageAnalysis function.
}
}
\value{
The coverage plot will be visualized.
}
\author{
Lukas Chavez
}
\examples{

print("This function has been deprecated. Please see MEDIPS.plotSeqCoverage instead.")

}
