\name{MEDIPS-package}
\alias{MEDIPS-package}
\alias{MEDIPS}
\docType{package}
\title{
MeDIP-Seq data analysis
}
\description{
MEDIPS was developed for analyzing data derived from methylated DNA immunoprecipitation (MeDIP) experiments followed by sequencing (MeDIP-Seq). 
	Nevertheless, functionalities like the quality controls may be applied to other types of sequencing data (e.g. ChIP-Seq). 
	MEDIPS adresses several aspects in the context of MeDIP-Seq data analysis.
}
\details{
\tabular{ll}{
Package: \tab MEDIPS\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2010-09-28\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.12.0), BSgenome\cr
}
}
\author{
Lukas Chavez, Joern Dietrich

Maintainer: Lukas Chavez <chavez@molgen.mpg.de>
}
\references{
Chavez et al., 2010
}
\keyword{ package }
